requirements.txt
instagrapi==1.12.11
opencv-python==4.5.3.56
Pillow==8.3.2
termcolor==1.1.0
discord.py==1.7.3
python-dotenv-0.19.0